% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRegression.R
\name{createLocalRegressionModel}
\alias{createLocalRegressionModel}
\title{Creates local regression model with normalized data}
\usage{
createLocalRegressionModel(dataset, normalizationTypes)
}
\arguments{
\item{dataset}{dataframe of training dataset; it will be normalized using \code{normalizeDataFrame} function}

\item{normalizationTypes}{vector of normalization type for each column.
Allowed values:
\itemize{
\item \code{minmax} - use min-max normalization
\item \code{zscore} - use Z-score normalization
\item \code{omit} - do not normalize
Character collumns will be normalized by default with \code{minmax} algorithm, normalization type value has no effect in that case
}}
}
\value{
Local regression model object
}
\description{
Creates local regression model with normalized data
}
