% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRegression.R
\name{localLinearWrap}
\alias{localLinearWrap}
\title{Perform local regression with \code{lm} regression model}
\usage{
localLinearWrap(localModel, test, n, formula)
}
\arguments{
\item{localModel}{LocalRegressionModel object builded with training data}

\item{test}{dataframe with test data for prediction}

\item{n}{number of neighbours for kNN algoritm}

\item{formula}{formula of predicted value}
}
\value{
dataframe of predicted values
}
\description{
\code{wrap} function using linear model as regression model.
Predicts values in test dataframe based on passed formula and LocalRegressionModel.
}
