% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRegression.R
\docType{class}
\name{LocalRegressionModel-class}
\alias{LocalRegressionModel-class}
\alias{LocalRegressionModel}
\title{LocalRegressionModel S4 class generator
Reserved for internal use, to create LocalRegressionModel use \code{createLocalRegressionModel} function}
\description{
LocalRegressionModel S4 class generator
Reserved for internal use, to create LocalRegressionModel use \code{createLocalRegressionModel} function
}
