% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRegression.R
\name{markClassifier}
\alias{markClassifier}
\title{Get errors of predictions}
\usage{
markClassifier(localModel, test, wrappedFunc, n, formula, ...)
}
\arguments{
\item{localModel}{model created with \code{createLocalRegressionModel} function}

\item{test}{dataset containing test data (along with the predicted column)}

\item{wrappedFunc}{prediction function that will be called - should return a vector of predictions}

\item{formula}{formula to be passed to \code{wrappedFunc}}

\item{...}{rest of the params passed to the \code{wrappedFunc}}
}
\value{
list of four element sqared mean error on test and train data
along with predictions of the \code{wrappedFunc} for the train and test data.
}
\description{
Mark classifier returning errors along with predictions on train and test datasets.
}
\examples{
classInfo <- .markClassifier(localModel, test, regressionTreeWrap, Sepal.Length~.)
}
