% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRegression.R
\name{normalizeDataFrame}
\alias{normalizeDataFrame}
\title{Normalizes dataframe for LocalRegressionModel}
\usage{
normalizeDataFrame(dataset, normalizationType)
}
\arguments{
\item{dataset}{dataframe with data to be normalized}

\item{normalizationType}{vector of normalization method for each column.
Allowed values:
\itemize{
\item \code{minmax} - use min-max normalization
\item \code{zscore} - use Z-score normalization
\item \code{omit} - do not normalize
Character collumns will be normalized by default with \code{minmax} algorithm, normalization type value has no effect in that case
}}
}
\value{
list containing:
\itemize{
\item \code{dataset} - normalized dataframe
\item \code{normalizationParams} - params used for normalization. Those values can be used to normalize vector outside dataset with same params.
}
}
\description{
Normalizes dataframe for LocalRegressionModel
}
