% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRegression.R
\name{normalizeEnumVector}
\alias{normalizeEnumVector}
\title{Converts vector of enum values (eg chr \link{"bmw", "bmw", "ferrari"}) to vector of normalized values in range \link{0,1}}
\usage{
normalizeEnumVector(v)
}
\arguments{
\item{v}{vector to be normalized}
}
\value{
list (dict) containing of \verb{$v} which is normalized vector and \verb{$map} which is list of mappings enum-to-assigned value
}
\description{
Converts vector of enum values (eg chr \link{"bmw", "bmw", "ferrari"}) to vector of normalized values in range \link{0,1}
}
\examples{
r <- normalizeEnum(c("a", "a", "b"))
# r$v = [0, 0, 1]
# r$map = list(a=0, b=1)
}
