% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRegression.R
\name{normalizeVector}
\alias{normalizeVector}
\title{Normalizes vector using provided \code{LocalRegressionModel}}
\usage{
normalizeVector(localModel, v)
}
\arguments{
\item{localModel}{LocalRegressionModel object}

\item{v}{vector to be normalised; it is required for vector to have same column order as training set specified in regression model.
Provided vector can be shorter than vectors in regression model.}
}
\value{
normalized vector
}
\description{
Normalizes vector using provided \code{LocalRegressionModel}
}
