% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRegression.R
\name{regressionTreeWrap}
\alias{regressionTreeWrap}
\title{\code{wrap} function using regression tree to build regression model.
Predicts values in test dataframe based on passed formula and LocalRegressionModel.}
\usage{
regressionTreeWrap(localModel, test, n, formula, method = "anova")
}
\arguments{
\item{localModel}{LocalRegressionModel object builded with training data}

\item{test}{dataframe with test data for prediction}

\item{n}{number of neighbours for kNN algoritm}

\item{formula}{formula of predicted value}

\item{method}{regression tree building method from \code{rpart} package}
}
\value{
dataframe of predicted values
}
\description{
\code{wrap} function using regression tree to build regression model.
Predicts values in test dataframe based on passed formula and LocalRegressionModel.
}
