% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRegression.R
\name{splitDataFrame}
\alias{splitDataFrame}
\title{Split dataset into training set and test set}
\usage{
splitDataFrame(dataframe, ratio, n = nrow(dataframe))
}
\arguments{
\item{dataframe}{dataframe to be splitted}

\item{ratio}{splitting ratio (\link{0, 1} value determining how much of dataset will be placed in train set)}

\item{n}{number of total rows in output dataframes}
}
\value{
list containing:
\itemize{
\item \verb{$train} train dataframe with $n * ratio$ rows
\item \verb{$test} test dataframe with $n * (1 - ratio)$ rows
}
}
\description{
Splits dataframe randomly into two dataframs at given ratio. You can provide total number of returned rows (chosen randomly form given set).
}
