% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRegression.R
\name{stripDependentVariable}
\alias{stripDependentVariable}
\title{Removes column with dependent value from dataframe based on formula}
\usage{
stripDependentVariable(dataset, formula)
}
\arguments{
\item{dataset}{dataframe}

\item{formula}{formula with variable names coresponding to dataframe}
}
\value{
list containing:
\itemize{
\item \code{dataset} - dataframe without column of dependent variable
\item \code{stripped} - removed column
}
}
\description{
Removes column with dependent value from dataframe based on formula
}
