% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRegression.R
\name{wrap}
\alias{wrap}
\title{Perform local regression with given regression model building algorithm}
\usage{
wrap(localModel, test, n, knnAlgorithm, func, formula, ...)
}
\arguments{
\item{localModel}{LocalRegressionModel object builded with training data}

\item{test}{dataframe with test data for prediction}

\item{n}{number of neighbours for kNN algoritm}

\item{knnAlgorithm}{algorithm used in kNN algorithm; allowed values: \code{kd_tree}, \code{cover_tree}, \code{CR}, \code{brute}}

\item{func}{function to be used in prediction}

\item{formula}{formula of predicted value}

\item{...}{additional params to be passed to \code{func}}
}
\value{
dataframe with prediction results
}
\description{
Predict values for \code{test} based on the \code{train} set with
function \code{func} calculating regression model on
some local data of length \code{n} provided by KNN-algorithm
The \code{func} will be called with \code{data} argument already provided
based on the \code{train} set. Any additional arguments can be
passed with \code{...} variadic argument.
}
\examples{
For usage examples look at implementation of
`localLinear.wrap` and `regressionTree.wrap` functions.
}
