% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRegression.R
\name{.wrapOne}
\alias{.wrapOne}
\title{Internal function allowing to perform wrapping on single observation}
\usage{
.wrapOne(localModel, x, n, knnAlgorithm, func, formula, ...)
}
\description{
Internal function allowing to perform wrapping on single observation
}
